Add Chef Workstation App installer feature for Windows
This adds a "Chef Workstation App" feature to the Windows
Installer.  It provides a 'run on login' component that will
add CWA to common startup

If Start menu items are selected, it will include this
shortcut there as well.

This branch also renames the shell shortcut "Chef Workstation" to "Chef
Workstation Command Prompt". This is more in line with recent
discussions around naming, and avoids confusion between a shortcut
called "Chef Workstation" and one named "Chef Workstation App".

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>